+++ branches/S0.7/src/textures.cpp
@@ -64,11 +64,13 @@ void TTexture::Draw() {
 	glGetTexLevelParameteriv(GL_TEXTURE_2D, 0, GL_TEXTURE_HEIGHT, &h);
 
 	glColor4f(1.0, 1.0, 1.0, 1.0);
+
+        // FIXME: Should this be GLint instead of all this casting of values to it?
 	const GLshort vtx[] = {
 		0, 0,
-		w, 0,
-		w, h,
-		0, h
+		static_cast<GLshort>(w), 0,
+		static_cast<GLshort>(w), static_cast<GLshort>(h),
+		0, static_cast<GLshort>(h)
 	};
 	glEnableClientState(GL_VERTEX_ARRAY);
 	glEnableClientState(GL_TEXTURE_COORD_ARRAY);
@@ -267,10 +269,14 @@ void CTexture::DrawNumChr(char c, int x, int y, int w, int h) {
 		texleft, 0
 	};
 	const GLfloat vtx[] = {
-		x,           Winsys.resolution.height - y - h,
-		x + w * 0.9, Winsys.resolution.height - y - h,
-		x + w * 0.9, Winsys.resolution.height - y,
-		x,           Winsys.resolution.height - y
+           static_cast<GLfloat>(x),
+           static_cast<GLfloat>(Winsys.resolution.height - y - h),
+           static_cast<GLfloat>(x + w * 0.9),
+           static_cast<GLfloat>(Winsys.resolution.height - y - h),
+           static_cast<GLfloat>(x + w * 0.9),
+           static_cast<GLfloat>(Winsys.resolution.height - y),
+           static_cast<GLfloat>(x),
+           static_cast<GLfloat>(Winsys.resolution.height - y)
 	};
 
 	glVertexPointer(2, GL_FLOAT, 0, vtx);
